users/urls.py
from django.urls import path
from rest_framework_simplejwt.views import TokenRefreshView

from users.apps import UsersConfig

from .views import UserCreateAPIView, UserTokenObtainPairView

app_name = UsersConfig.name

urlpatterns = [
    path("register/", UserCreateAPIView.as_view(), name="register"),
    path("login/", UserTokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
]
